Add `skip_in_debug_build`/`skip_on_ci` decorators
